feat(synthetix): perp refactor
