+++ src/apps/kwenta/optimism/kwenta.perp-v1-cross-margin.contract-position-fetcher.ts
@@ -3,9 +3,6 @@ import { gql } from 'graphql-request';
 import { ZERO_ADDRESS } from '~app-toolkit/constants/address';
 import { PositionTemplate } from '~app-toolkit/decorators/position-template.decorator';
 import { gqlFetch } from '~app-toolkit/helpers/the-graph.helper';
-import { GetDisplayPropsParams } from '~position/template/contract-position.template.types';
-
-import { SynthetixPerp } from '../../synthetix/contracts';
 import { OptimismSynthetixPerpV1ContractPositionFetcher } from '../../synthetix/optimism/synthetix.perp-v1.contract-position-fetcher';
 
 type GetCrossMarginAccounts = {
@@ -25,6 +22,7 @@ const getCrossMarginAccountsQuery = gql`
 @PositionTemplate()
 export class OptimismKwentaPerpV1CrossMarginContractPositionFetcher extends OptimismSynthetixPerpV1ContractPositionFetcher {
   groupLabel = 'PerpV1 cross-margin';
+  extraLabel = ' (v1 cross-margin)'
 
   async getAccountAddress(address: string): Promise<string> {
     const crossMarginAccountsFromSubgraph = await gqlFetch<GetCrossMarginAccounts>({
@@ -38,10 +36,5 @@ export class OptimismKwentaPerpV1CrossMarginContractPositionFetcher extends Opti
     }
 
     return crossMarginAccountsFromSubgraph.crossMarginAccounts[0].id;
-  }
-
-  async getLabel({ contractPosition }: GetDisplayPropsParams<SynthetixPerp>): Promise<string> {
-    const baseAsset = await this.getBaseAsset({ contractPosition });
-    return `${baseAsset}-PERP (v1 cross-margin)`;
   }
 }
