+++ src/apps/synthetix/optimism/synthetix.perp.contract-position-fetcher.ts
@@ -1,6 +1,7 @@
 import { Inject } from '@nestjs/common';
 import { parseBytes32String } from 'ethers/lib/utils';
 import { gql } from 'graphql-request';
+import { gqlFetch } from '~app-toolkit/helpers/the-graph.helper';
 
 import { APP_TOOLKIT, IAppToolkit } from '~app-toolkit/app-toolkit.interface';
 import { getAppAssetImage } from '~app-toolkit/helpers/presentation/image.present';
@@ -31,6 +32,7 @@ export const getContractsQuery = gql`
 `;
 
 export abstract class OptimismSynthetixPerpContractPositionFetcher extends ContractPositionTemplatePositionFetcher<SynthetixPerp> {
+  extraLabel = '';
   constructor(
     @Inject(APP_TOOLKIT) protected readonly appToolkit: IAppToolkit,
     @Inject(SynthetixContractFactory) protected readonly contractFactory: SynthetixContractFactory,
@@ -42,7 +44,24 @@ export abstract class OptimismSynthetixPerpContractPositionFetcher extends Contr
     return this.contractFactory.synthetixPerp({ address, network: this.network });
   }
 
-  abstract getDefinitions(): Promise<DefaultContractPositionDefinition[]>;
+  abstract marketFilter(market): boolean;
+
+  protected isV2Market(market): boolean {
+    const marketKeyString = parseBytes32String(market.marketKey);
+    //v2 marketKey includes 'PERP', v1 doesn't
+    return marketKeyString.includes('PERP');
+  }
+
+  async getDefinitions(): Promise<DefaultContractPositionDefinition[]> {
+    const contractsFromSubgraph = await gqlFetch<GetContracts>({
+      endpoint: 'https://api.thegraph.com/subgraphs/name/kwenta/optimism-perps',
+      query: getContractsQuery,
+    });
+
+    return contractsFromSubgraph.futuresMarkets
+      .filter(market => this.marketFilter(market))
+      .map(futuresMarket => ({ address: futuresMarket.id }));
+  }
 
   async getTokenDefinitions() {
     return [
@@ -54,7 +73,7 @@ export abstract class OptimismSynthetixPerpContractPositionFetcher extends Contr
     ];
   }
 
-  protected async getBaseAsset({ contractPosition }) {
+  private async getBaseAsset({ contractPosition }) {
     const multicall = this.appToolkit.getMulticall(this.network);
     const contract = multicall.wrap(this.getContract(contractPosition.address));
     const baseAssetRaw = await contract.baseAsset();
@@ -66,7 +85,10 @@ export abstract class OptimismSynthetixPerpContractPositionFetcher extends Contr
     return baseAsset;
   }
 
-  abstract getLabel({ contractPosition }: GetDisplayPropsParams<SynthetixPerp>): Promise<string>;
+  async getLabel({ contractPosition }: GetDisplayPropsParams<SynthetixPerp>): Promise<string> {
+    const baseAsset = await this.getBaseAsset({ contractPosition });
+    return `${baseAsset}-PERP${this.extraLabel}`;
+  }
 
   async getImages({ contractPosition }: GetDisplayPropsParams<SynthetixPerp>) {
     const baseAsset = await this.getBaseAsset({ contractPosition });
